api/requirements.txt
fastapi
uvicorn[standard]
qdrant-client[http]
sentence-transformers
python-dotenv
google-auth
openai
